move main.cpp code to game.cpp
